refactor: material 폴더명 변경 및 import 구문 변경
