+++ client/src/constants/weapon.ts
@@ -1,0 +1,28 @@
+export const weapon: Array<string> = [
+  "None",
+  "글러브",
+  "톤파",
+  "방망이",
+  "채찍",
+  "투척",
+  "암기",
+  "활",
+  "석궁",
+  "권총",
+  "돌격 소총",
+  "저격 소총",
+  "",
+  "망치",
+  "도끼",
+  "단검",
+  "양손검",
+  "폴암(미사용)",
+  "쌍검",
+  "창",
+  "쌍절곤",
+  "레이피어",
+  "기타",
+  "카메라",
+  "아르카나",
+  "VF의수",
+];
